Update ConvexHull2.h
The ConvexHull2 class was supposed to have switched from FPInterval to SWInterval, but I apparently overlooked this. The FPInterval class uses floating-point hardware operations to support interval arithmetic; however, the GCC compiles ignores the changes to the floating-point environment (it does not support modifying the rounding mode of the FPU). The SWInterval class implements interval arithmetic and rounds using software operations.