Feature request: Modifications related to rotating calipers.
Added an implementation of the rotating calipers algorithm, class RotatingCalipers. The goal was to refactor MinimumAreaBox2 to pull out the rotating calipers. However, this does not work because the minimum-area algorithm uses a pair of rotating calipers that work together. Added a new class MinimumWidthPoints2 that uses the rotating calipers algorithm to compute the width of a 2D point set. For now I have modified the MinimumAreaBox2D sample application to show also the results from computing the width of a point set. I also cleaned up the comments to use current formatting rules. And I added functions to allow passing std::vector inputs rather than raw pointers.