Ported WM5 planar reflection sample to GTE6.
I wrote a GTE6 sample application for planar reflections that is similar to the one for Wild Magic 5. The PlanarReflectionEffect drawing function had a bug. It needed to set the back-face stencil parameters in addition to the front-face stencil parameters. Without the back-face parameters, the reflection caster was drawn on both sides of the reflection plane.