Update IntrSegment2Segment2.h
The current 2D segment-segment queries use the find-intersection query for two lines, but the segments are converted to center-direction-extent form. The two-point representation has endpoints p0 and p1 and the parameterization is p0 + t  (p1 - p0) for t in [0,1]. The center-direction-extent form is c + s d for |s| <= e. The center is c = (p0+p1)/2, the direction is d = (p1-p0)/|p1-p0| and the extent is e = |p1-p0|/2. The computation of d involves a normalization which generally cannot be computed exactly with rational arithmetic. The computation of e is also not exact. The segment parameters returned from the queries are relative to s, not to t. The old queries are operator() functions and still exist. To allow for exact rational computation, I added new queries named Exact which use the two-point form for segments. [In GTL development, I have been eliminating the use of the segment center-direction-extent form.]