Modification for MS Windows to support third-party UI package.
By user request, on Microsoft Windows I added virtual OnWindowsMessage to WindowApplication and called by WindowSystem::WindowProcedure. This allows the application to use a third-party UI package, processing windows messages intended for that package. A return value of true tells the window procedure not to pass the message to the application window.