Fixed a bug in Transform::Inverse.
In the GTL development, fixed a bug in the Inverse function for Transform<T> objects. The unit tests were incomplete in that they tested the inverse directly using queries to the internal matrices of Transform<T> rather than verifying the correctness of the output of Inverse. The corrections were backported to GTE.