Update NaturalSplineCurve.h
The constructor for free or closed splines allocated more than enough elements for mCoefficients and the GetNumPoints() function returned the number of points based on the size of mCoefficients. The constructor for clamped splines allocated exactly the correct number of elements for mCoefficients, but then GetNumPoints() returned the incorrect number of points. This caused too few points to be processed in clamped splines. This occurred when the optimization was added to minimize memory allocations and deallocations for free splines. The optimization code has been reworked and the same memory optimization was added for closed and clamped splines. The unit tests pass for all 3 flavors of splines.