Update UniqueVerticesTriangles.h
Fixed an error in the deprecated file for UniqueVerticesTriangles; the <array> header is not explicitly included. This occurred with gcc 12.1.1 on Fedora 36 even though my instantiation tool did not include explicit instantiation for that class. The error did not occur with gcc 9.3.0 on Ubuntu 20.04.1 LTS or Microsoft compilers when running the tool.