Fixed gcc 12.1.1 warning about potential index-out-of-range.
Fedora 36 with gcc 12.1.1 complains about potential index-out-of-range because
Evaluate references all possible elements, but the compiler is unaware that the logic of Evaluate is designed to assign only those elements that are valid. For now, I am passing in an array of 3-tuples knowing that only one will be filled in. Added arrays of  N-tuples to avoid the warnings.