Create VETManifoldMeshNR.h
Added a new class VETManifoldMeshNR which represents a vertex-edge-triangle manifold mesh for which triangles are provided as a single batch and NO REMOVE operations are going to be performed on the mesh. VETManifoldMeshNR significantly outperforms VETManifoldMesh. VETManifoldMesh is general purpose, allowing triangle insertions and triangle removals. The underlying C++ container classes lead to significant memory allocation/deallocation costs and are also expensive for find operations. It turns out that the design of VETManifoldMeshNR automatically gives you vertex adjacency information, so there is no ETManifoldMeshNR implementation. It is a requirement that the input triangles form a manifold mesh with consistently ordered triangles. In most applications, this
requirement is already satisfied.

The VETManifoldMeshNR class is designed to minimize the allocations. For example, using this class for a collection of 1131652 positions and 2242293 triangles, the CPU times on an Intel (R) Core (TM) i9-10900 are as follows. For ETManifoldMesh, graph creation used 2234 milliseconds, connected component finding used 2169 milliseconds, and finding the boundary polygons used 232 milliseconds. For VETManifoldMesh, graph creation used 118 milliseconds, connected component finding used 40 milliseconds, and finding the boundary polygons used 25 milliseconds.