Added header path to managed library sample.
The managed library project compiles with the directory structure of GTE, but if folders are moved elsewhere, the managed library needed $(GTE_PATH) for the header search path. GTE_PATH is an environment variable that stores the path to the top-level folder GeometricTools/GTE (see the installation and release notes).