Update IntrEllipse2Ellipse2.h
When using BSRational<UIntegerAP32> for template type T, the number of words used in the SolveQuartic root finder can be extremely large---on the order of 100000. This leads to poor performance. The root finder has numerical rounding errors whether T is a floating-point type or a rational type, so to avoid the poor performance, the rational-valued quartic coefficients are converted to double precision and then reconverted to T. This introduces a small amount of rounding
error, but performance is significantly better. I also added initialization of Result members in the constructor to avoid compiler warnings.