Update IntrEllipse2Ellipse2.h
Rolling back the changes from the 4 November 2002 update. The code is faster but the rounding errors cause misclassification of the number of points of intersection. Welcome to the trade-off of floating-point versus rational arithmetic. At some point I need to implement rational arithmetic classes using the GNU Multiprecision Library to take advantage of the better performance for arbitrary precision multiplication. At a later date, I can implement a similar speed-up for BSNumber and BSRational.