Clean up after the previous merge.
Modified the "File Version" numbers to 2023.02.02. The default constructor for Window::Parameters in Window.cpp needed to set the new member 'parent' to nullptr. The new callback OnCopyData in WindowApplication.{h,cpp} exposes Win32 data types. The new code needed to be in conditional blocks enabled when GTE_USE_MSWINDOWS is defined to avoid breaking the Linux build. Modified the update PDFs with information about the changes.