New OBB tree code (tree for points or segments or triangles).
Redesigned the OBB tree of points, factoring out the tree construction to
\Code{OBBTree.h}. I also fixed an incorrect sizing of the nodes of the tree.
A feature request was made of an OBB tree of triangles, which was easy to
implement by derivation from \Code{OBBTree}. The \Code{OBBTreeOfTriangles}
class has additional support for computing the intersection of linear
component (line, ray, segment) with a triangle mesh. I also added code for
computing an OBB of segments.