Update RootsPolynomial.h
Rewrote the biquadratic implementation. The old code had a problem in one computation where subtractive cancellation caused many significant bits to be eliminated. This was the problem in the github issue about intersection of two ellipses, one rotated by pi/4+epsilon for a small epsilon. An algebraic identity was used to avoid the subtractive cancellation, and the ellipse-ellipse intersection problem was eliminated. The problem with ellipse-ellipse intersection for rotation epsilon still exists when you are using floating-point arithmetic. It does not occur with rational arithmetic, but I am rewriting that code to fix the floating-point problems).