Fixes to line-distance query in 3D
DistLine3Circle3.h: The line-circle distance code had flawed logic for the case of a line not perpendicular to the plane of the circle and containing the origin. However, the block is entered when the line origin is on the normal line to the plane containing the circle center.

RootsPolynomial.h: Two std::sqrt calls have arguments that are theoretically nonnegative, but floating-point rounding errors can lead to negative arguments. I added guards against this (clamp to nonnegative). The other \Code{std::sqrt} calls already have the guard. The errant calls were in SolveDepressedQuartic. The problem was discovered during the line-circle distance query investigation.

Update the history PDFs.