Added new file RootsQuadratic.h and modified rational number files.
Added a new class RootsQuadratic for robust estimation of roots to quadratic polynomials. The robust versions for cubic and quartic polynomials use this; the work is in progress. The older code in RootsPolynomial will remain in GTE for now but will be deprecated in GTL.

Added number.Negate() member functions BSNumber and BSRational to avoid the verbose number.SetSign(-number.GetSign(). BSRational.SetSign had an mValue block with GTL_BINARY_SCIENTIFIC_SHOW_DOUBLE instead of GTE_BINARY_SCIENTIFIC_SHOW_DOUBLE. This affected only the visualizer display of mValue; the rational number itself is correct.