Replaced initial cone estimation by simpler algorithm.
Modified the code to initial cone parameters when the caller requests initialization. The previous initialization was an implementation of an algorithm described in the LeastSquaresFitting.pdf file but the mathematics of the file was incorrect. Fixing the mathematics leads to having to compute roots of a degree 10 polynomial. Instead, a simpler method is provided that uses a line-fitting algorithm to determine the relationship between height h and radial distance r. The PDF itself has been modified.