Update to ellipse-ellipse intersection.
Modified the ellipse-ellipse find-intersection query code to be more robust when using floating-point arithmetic. The code still executes with rational arithmetic. Added a sample program for testing the code. The interface for FIQuery<T,Ellipse2<T>,Ellipse2<T>} changed, so the area-of-intersection code had an obsolete interface removed. Added support for fused-multiply-add instructions.