Choose DX11 discrete adapter before Intel Integrated Graphics
Running a ConsoleApplication-derived object on a laptop with Intel Integrated Graphics (Iris Xe/UHD) and an NVIDIA GeForce RTX 3050, the Intel graphics is identified  as the most powerful adapter because that adapter is associated with the primary display. I had added code to WindowApplication-derived objects some time ago to choose a discrete adapter (if available) before choosing the Intel driver. That same code has now been added to the console-based application layer.