Fixed bug in DistLine3Circle3.
The code ported to GTL had a unit-test failure for the Robust() function in the case that the circle center was not the origin. The GTE unit tests did not contain such a case. One block of Robust() was missing the translation of the closest line point to the original coordinate system.