Update DistLine3Circle3.h
This file is cursed. Last check-in, I merged a line from GTL to GTE. The query result is stored in a struct Output in GTL but the GTE query result is stored in a struct Result. I then failed to run my instantiation tools which would have trapped the error.