Restructuring header files.
A large number of files have been modified in order to ensure that each header file includes any C++ Standard Library headers it references. This avoids breaking the compile when GTE headers are refactored. I also unified the commenting so that top-level comments occur after the pragma-once statement but before any header file includes. (All of this had already been done in GTL and ensures that the instantiation tool behaves correctly.) See Gte6UpdateHistory.pdf for the details of all the changes.