Fixed bugs in point-line and point-ray queries.
The comments in these files indicate the direction vectors do not have to be unit length. This was not true in earlier versions of GTE; that is, the directions had to be unit length. I had modified the comments at one time because I was removing the unit-length constraint. Unfortunately, the code was not modified, so only unit-length vectors worked. My unit tests passed anyway because they all used unit-length directions. The constraint is now removed. The direction vectors do not have to be unit length, the code was corrected, and the unit tests were updated to handle non-unit-length vectors. The line-segment distance query in DistPointSegment.h is correct.