Fixed problems in SurfaceExtractorMC.h
When the single-voxel Extract returns false, the full-image extraction was supposed to ignore the voxel. Instead it deleted the entire mesh, which is a disastrous bug. The unit tests verified only that the Marching Cubes 256 triangle-lookup table entries are correct, and none of these provided code coverage that would have found the bug. Also, floating-point rounding errors in mesh vertex construction led to small sets of vertices, each set containing vertices that are nearly the same but should have been the same without the rounding errors. MakeUnique is unable to identify these as duplicates. I modified the single-voxel Extract function to generate the same vertex on an edge shared by multiple voxels, using a consistent ordering of the voxel corners to ensure that the all constructions of the vertex have the same rounding errors. Finally, the interfaces were modified to have a new parameter 'perturb'
that allows you to avoid the ignoring of voxels for which at least one corner of the voxel is exactly the input 'level' value. Thanks to Scott Johnson for reporting these problems.