Additional modifications to SurfaceExtractorMC.h.
In SurfaceExtractorMC.h, the test code used only a positive perturbation and the output is correct. When the perturbation is negative, the order of sign testing needed to be modified to ensure the output is correct. The BouncingBall physics sample uses
SurfaceExtractorMC.h, so it needed a modification to add the perturbation parameter.