Fixed a bug in IntrLine2Circle2.h, removed comment in Arc2.h
Fixed a problem with segment-arc and ray-arc find-intersection queries that were caused by the line-circle find-intersection query setting only one of the result.parameter values when the line and circle are tangent. It is necessary to set both parameter values to the same number, effectively a degenerate interval that is used in interval-interval find-intersection queries used in segment-arc and ray-arc. Thanks to owai on github for reporting the problem, diagnosing it, and fixing it.

Removed the deprecation comment about the Arc2 Contains function without the epsilon argument. This function is still useful when you know the point is on the arc so that you do not need the epsilon value.