Deprecated RootsPolynomial.
The RootsPolynomial<T>::Find function suffered from significant floating-point rounding errors in the computation of the coefficients of the polynomial derivatives. The class allows for type T to be rational, but the bisection effectively does not converge because the number of bits per rational number is excessively large. This class is now deprecated. The low-degree root finders are now in RootsLinear.h, RootsQuadratic.h, RootsCubic.h, and RootsQuartic.h. Three of these files are listed here because a comment mentioned using Cauchy's bound when in fact the code
uses Lagrange's bound.