Replacement for RootsPolynomial::Find.
The replacement for RootsPolynomial<T>::Find is RootsGeneralPolynomial<T>::Solve. The algorithm is effectively the same as the deprecated Find, but it uses a mixture of floating-point arithmetic and rational arithmetic. The coefficients of the polynomial derivatives are computed using rational arithmetic to avoid the rounding errors that plagued the deprecated code. When the bisection code consumes these polynomials, rational evaluation is used to determine sign information at the endpoints of an interval. If bisection must occur, the rational coefficients are converted to floating-point, and the resulting rounding errors are significantly smaller than that of the deprecated code. The floating-point-based bisection is fast, so the performance is reasonable in order to get increased accuracy of the root estimates.