Update ETManifoldMesh.h
Added an exception to the Insert function that detects whether the vertex ordering for two edges is opposite for the two triangles sharing the edge. If the ordering is not opposite, the mesh is not manifold. The test is exposed by default (mThrowOnNonmanifoldInsertion is initially true), but you can disable the test by setting the aforementioned Boolean to false.