Update SurfaceExtractorMC.h
The GetGradient function had a typecast of a float to a size_t and then assigned it to an int32_t. The typecast was changed to an int32_t. Fixed Ubuntu gcc compiler warnings/errors: missing typename when the member topology, 3 loop variables were size_t but needed to be IndexType.