Update IEEEBinary16.h
LLVM - clang-cl error. An include of Functions.h is required; otherwise, the block of functions starting with atandivpi become infinitely recursive. The definition of atandivpi whose input is IEEEBinary16 makes a call to atandivpi whose argument is typecast to float. This works as long as the definition of atandivpi in Functions.h is accessible. Without it, an implicit IEEEBinary16 constructor call is made from the float input, thus the infinite recursion. The constructor with input of type double has variable temp that is tagged as not initialized. It is on the next line, but I moved that code into the union initialization.