Update DistCircle3Circle3.h
Reimplemented the query for distance between two circles in 3D. The previous version used only floating-point arithmetic. The rounding errors when computing derivatives of the degree-8 polynomial were causing extreme inaccuracy in the results. The new code now uses a mixture of floating-point arithmetic and rational arithmetic. The degree-8 polynomial and its derivatives are created using rational arithmetic. Bisection is performed using floating-point arithmetic, but with some initial rational arithmetic to determine sign changes in a polynomial at the endpoints of a potential root-bounding interval.