Fixed a bug in the nonmanifold test in the Insert function.
The nonmanifold test for a shared edge having opposite orderings for the two triangles sharing the edge was flawed. The 'edge' object comes from an EdgeKey<false> object so that the edge is <v0, v1> with v0 < v1. The actual ordering depends on the specified triangle vertex order. Instead, the comparison has to be made for the shared edge of the already existing triangle edge->T[0] because the vertices for that triangle are ordered.