Added ability to extract filaments.
The comments indicated that the filaments are extracted by the code. They are removed from the graph but not made available to the user. I added the ability to extract the filaments in case the caller wants them. The file is relatively old, so I took the opportunity to upgrade it to my current design goals. The constructor is a simple one, no longer mimicking a function object. The Extract function does the cycle and filament extraction. I also added a new template parameter which is the type of the vertex indices. The type was int32_t but now can be any signed or unsigned integer type of size at least 2 bytes. Thanks to Scott Johnson for (once again) reporting problems such as these.