Update TriangulateEC.h
Fixed bugs that were discovered when triangulating a polygon tree. The Vertex objects of the VertexList had state information from the previous triangulation that affected DoEarClipping. The state is now reset to the constructor values in the Vertex constructor. The polygon-tree triangulation also was not accumulating the triangles for the multiple DoEarClipping calls.