Another revision of MinimalCycleBasis.h.
The new implementation had multiple problems. The code was complicated enough that more care was needed to get it right. My previous attempts were a poorly handed response to a request for a new feature. I stepped through the code for the latest version using a debugger, specifically for SimpleGraph0.txt of the sample application. The code coverage hits all the relevant blocks. The unit tests for the simpler configurations pass.