Update RoughPlaneParticle1Window2.cpp
The application class member mContinueSolving was uninitialized. The debug build worked for me but not the release build. Added class member initialization to the application constructor. When SINGLE_STEP is enabled, the code does not compile because of a porting error. It had a Wild Magic Vector2d exposed but needed to be a GTE Vector2<double>.