Posting GTE 7.0.
The README.md has a link to the full update history which contains the changes made to post version 7.0. I have modified the copyright notices for all the source files, but the file version numbers remain the same.

The Ubuntu and Fedora tools were updated to latest (gcc/g++, CMake, Visual Studio Code) and the sample applications were all compiled and executed to ensure they work.