Update ExtractRidgesConsole.cpp
The finite-difference approximation to the second-order derivative f_yy had a typographical error. Also, to avoid sign changes artificially generated because the numerical eigensolver can be inconsistent in returning eigenvectors: At one point an eigenvector U is returned but at a nearby point an eigenvector is return that is approximately -U. Thanks to Lev A. Melnikovsky for reporting the problems and fixing them.

My changes include updating the file version in the header comments and modifying the new code to use my coding standards.