Update InscribedFixedAspectRectInQuad.h
Fixed minor issues after porting the code and unit tests to GTL: Removed test code. The isUnique variable was set to false when the intersection of intervals is a nondegenerate interval. Instead, isUnique must be set to true when the w-values are the same at the interval endpoints, false otherwise.