Updated the Hermite interpolation code.
Ported the Hermite interpolations from GTE to GTL. Verified the systems of equations and solutions that produce the coefficients of the polynomials, both using code that generates code for the system and by using Mathematica to solve symbolically the numerous cases. Unit tests for the GTL code are more rigorous than those for GTE, so the GTL code and unit tests were back-ported to GTE. The SmoothLatticeInterpolation.pdf document has been updated significantly to describe the coefficient construction.