Fixed template instantiation problems.
My template instantiation system needed serious modification for when I deprecated a class and provided a new implementation. The system failed to instantiate the new implementations. I discovered this when attempting to backport the IntpQuadraticNonuniform2 class (the Cendes-Wong algorithm) from GTL to GTE. The GTL code has been thoroughly unit tested and verified with Mathematica. There are no sample applications that exercise those new implementations, which was clear when the code would not compile---a serious oversight on my part. I have made changes to various classes to get the template instantiation to compile the new code. I will post the backported GTL code for IntpQuadraticNonuniform2 after I unit test the derived classes IntpSphere2 and IntpVectorField2.