Bug fix to Interpolation2D sample application.
The classes IntpQuadraticNonuniform2 for GTE and GTL pass the same unit tests, and those results were verified by Mathematica for the GTL class. The rendering in the interpolation sample application clearly showed the derivatives did not match at the boundaries of the center triangle of the 4 triangles. It turns out that the sample code itself had a bug. The SampleMesh class initializes mAdjacencies to have the triangle indices for neighbors. For the triangle V1,V4,V3, the adjacencies were assigned as 0, 1, 2. The assignment needed to be 1, 2, 0. After this modification, the rendered surface is visually smooth. The window client size was 512x512, but I made it larger for a better view of the renderings. It is now 768x768.