Update LieGroupsAlgebras.h
Added a new PDF (LieGroupsAlgebras.pdf) that describe the mathematics and implementation of LieGroupsAlgebras.h. The minimax approximations had to be adjusted for the cases when |theta| > pi. Also, the adjoint representations were those of Ethan Eade's document (mentioned in the bibliography of the PDF), but they were based on his choice for the order of the generators. My implementation uses a different order, which leads to adjoint representations that are permutations of his.