Modifications for the last pull request.
The merge of a pull request added a second pragma-once statement to ContTetrahedron3.h, which I removed. The pull request also fixed typographical errors in the comments, and a Python script was added to removed files from the GTE directory tree. The TrackObject class was given setters for x-size and y-size that are used if the application window is resized. I updated the file version numbers for these modifications.