Update IntpAkimaUniform3.h
Fixed bugs in the interpolator. The original Akima code was in Wild Magic 5, and it was ported correctly to GTE 1.0. The incorrectly repeated FX calls were introduced in GTE 2.2 (2/2/2016). The lack of parentheses for (bound-1) expressions was introduced in GTE 5.14 (11/11/2021). The Akima uniform 3D interpolator had been unit tested in Wild Magic, but I dropped the ball on testing it during the evolution of the GTE code. Thanks to Russel Wenzel for reporting the errors.