Fixed subtle bug in BSNumber debugging code.
Unit testing of TriangulateCDT in the GTL development track exposed a subtle bug in BSNumber when GTE_BINARY_SCIENTIFIC_SHOW_DOUBLE is defined. The functions BSNumber::SetSign, BSNumber::SetBiasedExponent, and BSNumber::SetExponent each recompute mValue. The problem is that if the source object represents 0 and SetSign is called with a nonzero value, the target object is not yet a valid BSNumber because SetBiasedExponent and bit copies have not yet occurred. This leads to failure in the mValue recomputing of SetSign. The GTE version of TriangulateCDT does not fail because 0 is never the source of the
incremental construction, whereas the GTL version has a case where the source is 0. GTE has several blocks of code that incrementally create a BSNumber object using SetSign, SetBiasedExponent, and a copy of the integer bits from the source object. I added updates of mValue after those blocks.