Update MinimumVolumeBox3.h
The GetExtreme function was incorrect when all the adjacent vertices of the initial vertex have the same projection in the specified direction. Geometrically, the convex hull has a collection of coplanar triangles sharing the initial vertex. GetExtreme bails after examining the adjacent vertices and incorrectly concludes that the initial vertex is the extreme. The comparison (dCandidate > dMax) must instead be (dCandidate >= dMax) to allow the hill climbing to continue away from the flat region of the initial vertex. And it is now necessary to use visited flags to ensure the hill climbing does not oscillate between two vertices of the same projection value. The visited flag array must be local to GetExtreme because in the multithreaded case, a class member array is no longer protected from concurrent access. Thanks to github's AnsonInTheAstroworld for reporting the bug.