Update ImageUtility2.h
When drawing a 2D disk using ImageUtility2::DrawCircle with a Boolean solid of true, for each x-value the callbacks are executed for a range of y-values. For a callback as simple as drawing only the pixel, this leads to many cache misses because of the row-major ordering for the image. The code has been modified to select y-values and execute the callbacks for a range of x-values.