Update ConvexHull2.h
The convex hull is computed using rational arithmetic. The class IntrinsicsVector2 is first used to determine the intrinsic dimension of the input points, but these computations use floating-point arithmetic. The direction between the extreme points is obtained as a normalized difference of points. The floating-point rounding errors can lead to the incorrect conclusion that the hull is 2-dimensional, but the hull (using rational arithmetic) is actually 1-dimensional. The dataset that exposed the bug was passed to the minimum-area box code, and the misclassification of the
intrinsic dimension leads to a crash caused by an out-of-range index into a std::vector. The dataset has more than 3 double-precision points that happened to be exactly on a line; rational arithmetic verified this. If exactly 3 double-precision and colinear points are passed to the hull finder, there is no crash. I removed the IntrinsicsVector2 code and set mDimension based on the rational computation of the hull. As it turns out, I had already done the same in the GTL code but for some reason did not back-port that code to GTE. Thanks to Richard Bouwman for reporting the bug.