Update to convex hull 2D and dependent code.
Removed the obsolete mEpsilon member, which changes the signature to the operator() functions. Added an operator() function whose input is a std::vector<Vector2<T>>. Back-ported (most of) the GTL ConvexHull2 code to GTE. Fixed compiler breaks that resulted from the signature change. As is the case with other reworked GTE files, I replaced the template parameter Real with T as a signal that I have revisited the code (usually for GTL and back-porting to GTE).