Update MinimumVolumeBox3.h
The GetExtreme function still did not work correctly after the July 24, 2024 modification. The initial value of dMax for vertex mVertices[0] is computed before the loop over the vertices to find potentially larger dMax values. An example showed that the initial dMax is 0. The adjacent vertices to mVertices[0] all have zero-valued d-values. The adjacent vertices to mVertices[1] all have zero-valued d-values, and one of those vertices is mVertices[0]. This causes vMax to be assigned to 0 again, which invalidates the search for maximum dMax. I added another line before the loop that marks mVertices[0] as visited. Thanks to github's AnsonInTheAstroworld for reporting the bug and providing an example that caused failure.