Another fix for GetExtreme in MinimumVolumeBox3.h.
One more time. The GetExtreme function still failed. The essence of the function is a hill-climbing algorithm using a queue. My poorly written queue-like code, and in particular the visited[] management, was still not correct. Thanks to github's AnsonInTheAstroworld for reporting the bug once again, and for providing a dataset for which the implementation failed. Although a modification was suggested using queues (with storage std::vector), I used the simplest algorithm, which is to iterate over all the vertices and compute the maximum height. I used Intel oneAPI VTune for profiling. It turns out that the bottleneck in computing are the memset calls in the constructors for mBits of UIntegerFP32<2561>}.
MinimumVolumeBox3 has a significantly large number of constructor calls; the initialization was approximately 80% of CPU time (the latter number on the order of  1 minute). I removed the initializations. On my Intel i9-10900 CPU which has 20 logical processors, the profiling showed that with a
single-threaded execution (in the main thread), the queue-based approach used 4638 milliseconds for the provided dataset. My simple algorithm used 3121 milliseconds. Using 8 threads, the queue-based implementation used 1028 milliseconds and my simple algorithm used 722 milliseconds. The numbers are
sufficiently small enough that the simple algorithm should suffice.