Modifications for parallelogram2 and distance query.
GTE/Mathematics/DistPoint2Parallelogram2.h: The inclusion of the header file for Parallelogram2.h was exposed via quotes but needed to be <Mathematics/Parallelogram2.h>. I also added includes of C++ Standard Library headers whose symbols are accessed in the distance query header.

GTE/Mathematics/Parallelogram2.h: Modified the order of vertices returned by GetVertices so that the ordering is consistent with that of OrientedBox2.