Update UIntegerFP32.h
The Microsoft Visual Studio 2022 (version 17.12.0) code analysis tool complained about not initializing mBit in the constructor initialization list. I had added a note to the file that initialization leads to a severe performance penalty (setting large numbers of array elements to 0). I added a pragma to disable the warning for the constructors.