Update WGLExtensions.cpp
The Intel C++ Compiler 2025 complained about the static cast of the function pointer returned from wglGetProcAddress. I removed the static cast and used the already existing reinterpret_cast in GetOpenGLFunctionPointer from the return pointer to a void* followed by a reinterpret_cast} to the appropriate function pointer type. The compiler also warns that GetOpenGLFunctionPointer can be made
static, but that is incorrect. This function is implemented and exposed via extern in the WGL engine and in the GLX engine.