Rewrite to fix bugs in the intersection of cylinders.
Equation (16) in IntersectionOfCylinders.pdf has term |Delta|, but it should be |Dot(D(x,y), Delta|$. The implementation used the incorrect term. I removed the complicated argument for selecting potential separating directions from the pole plane, replacing it with selecting potential separating directions directly from the hemisphere. The TIQuery constructor now requires specifying the number of theta and phi angles for spherical coordinates on the hemisphere. It also has an input for selecting the number of threads to use for the separation tests. If the input is 0 or 1, the selection is performed in the main thread of the application. If the input is positive, the application runs multithreaded. The threading is designed so that if a separation axis is found by one of the threads, all the threads terminate. Thanks to github's Gerhard Wesp for reporting the bug.