Update IntrCylinder3Cylinder3.h
After revising the PDF, I noticed that the multithreading code process j = 0, which corresponds to phi = 0. This is the north pole N of the hemisphere, so numTheta separation tests occur for N. I modified jmin.first() to be 1, not 0. I also noticed the phi values were sampled on [0,pi]. For a hemisphere, they only need to be
sampled on [0,pi/2]. The revised PDF is now uploaded to the Geometric Tools
website documentation folder.