Matrix preprocessor blocks of code reorganized.
The installation notes mentioned specifying project preprocessor symbols
GTE_ROW_MAJOR or GTE_COL_MAJOR and GTE_USE_MAT_VEC or GTE_USE_VEC_MAT. If a user forgets to define matrix storage order or matrix-vector multiplication convention, some GTE code can have incorrect behavior. This is a result of inconsistently choosing a preprocessor symbol to control the branching of conditionally compiled code. GTE by default uses row-major order and matrix-vector products using the vector-on-the-right convention. I modified the files so that branching is based only on GTE_COL_MAJOR and GTE_USE_VEC_MAT. If the matrix storage order and multiplication conventions are not defined using
the preprocessor symbols, the code will compile for row-major order and
vector-on-the-right convention.