Update DistCircle3Circle3.h
The inverse transform of the closest points had a bug. The point for circle1 was not inverted. I had also added code to order the circles internally so that query(circle0,circle1) and query(circle1,circle0) return the same value. I had fixed the
bug in GTL and added the swap in GTL but failed to backport the code to GTE. Thanks to zhou xun for reporting the errors.