Added specialized constuctors to Vector.
I added specialized constructions via a std::enable_if mechanism for dimensions 2, 3, and 4. This reduces constructor time for those dimensions
by half compared to the std::initializer_list constructor. The latter constructor is now specialized for dimensions larger than 4. In doing so, I also added a constructor that takes a single value and assigns it to all components. Implicitly, Vector<N,T> zero{(T)0} sets the vector to the zero vector. Fixed compile breaks for code that used the std::initializer_list constructor but did not provide a full list of components. (I noticed the performance problem when profiling find-intersection code with a large number of calls, causing member initialization of the Result struct to be a bottleneck.)