Added new code, fixed minor version in CMakeLists.txt files
Added a new class PolygonWindingOrder that determines whether a 2D polygon is counterclockwise or clockwise ordered in O(n) time. Added a new class for computing the convex hull of a simple polygon in O(n) time. Added a sample application as a unit test. I had forgotten to update the minor version to 3 (GTE is now 7.3) in the CMakeLists.txt.